Postgres client - link to /usr/local/bin

+++ scripts/tools/postgres.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 
 # postgres client (link needed as by default psql is not linked to /usr/local/bin)
-brew install libpq  
-brew link --force libpq  
+if [ ! -f /usr/local/bin/psql ]
+then
+    brew install libpq
+    brew link --force libpq
+fi
 
 # pgadmin4
 brew cask install pgadmin4
